==> pipelines/4__indexing.ipynb <==
# <b>Created DENSE and SPARSE indexes through various indexing techniques.</b>

#### 1) Dense index (Semantic) with FAISS.

FAISS is perfect for 50K - 500K vectors on one machine.

import numpy as np, faiss
from pathlib import Path
import os, math, time, numpy as np, pandas as pd, torch

from elasticsearch import Elasticsearch, helpers

MODEL_TAG = "paraphrase-multilingual-MiniLM-L12-v2"
METADATA_FILENAME = f"{MODEL_TAG}__meta.parquet"
METADATA_FILENAME

df_meta = pd.read_parquet(f"data/embed/{METADATA_FILENAME}")  # includes global_chunk_id, etc.

df_meta.head()

Load metadata and embedding files.

EMBED_FILENAME = f"{MODEL_TAG}_384d_50k_float32.npy"
EMBED_FILENAME

# Load your embeddings and metadata (from earlier step)
embeddings = np.load(f"data/embed/{EMBED_FILENAME}").astype("float32")  # (N, d) L2-normalized


# Validate dimensions.
dim = embeddings.shape[1]

## Write index to FAISS.

## Filename.
INDEX_FILENAME = f"{MODEL_TAG}__indexes.faiss"

INDEX_FILENAME

# use IP since vectors are L2-normalized -> cosine
index = faiss.IndexFlatIP(dim)

# Add embeddings.
index.add(embeddings)

# Write index to the shared data library.
faiss.write_index(index, f"../shared-data-library/indexes/faiss/{INDEX_FILENAME}")

****

## <b>Keyword based indexes (SPARSE)</b>

### <b>2. Prepare data & bulk-ingest passages</b>

# Example: load the duplicates file
df_passages = pd.read_parquet("../shared-data-library/out/df_passages.parquet")

# Ensure stable IDs
df_passages = df_passages.copy()
df_passages["global_chunk_id"] = (
    df_passages["doc_id"].astype(str) + ":" + df_passages["chunk_id"].astype(int).astype(str)
)

# Minimal columns for ES
to_index = df_passages.rename(columns={"site": "site"})[
    ["global_chunk_id","doc_id","chunk_id","site","lang","title","chunk_text","preview","chunk_tokens"]
].fillna({"title":"", "preview":"", "chunk_text":""})

# Create instance of elasticsearch server.
es = Elasticsearch(
    "http://localhost:9200",
    basic_auth=("elastic", "DfZP9TzO")   # 👈 add this
)


def gen_actions(df):
    for r in df.itertuples(index=False):
        yield {
            "_index": "passages_bm25",
            "_id": r.global_chunk_id,
            "_source": r._asdict()
        }

# Ingest
helpers.bulk(es, gen_actions(to_index), request_timeout=180)

# Refresh sparse index.
es.indices.refresh(index="passages_bm25")

******
